Detect nightly build status
The build in branch tb-mingw, which is used on Windows, has been failing.
This has caused a 500 error for those who are trying to download the
package.

I've decided to direct the user the the build status page and asked them
to consider writing a new issue on Github against the neovim project.